AOD: check in some changes laying around

+++ schemas/+ephys/SpikesAlignedTrial.m
@@ -28,8 +28,8 @@ classdef SpikesAlignedTrial < dj.Relvar
             tuples = dj.struct.join(key,trials);
             
             for tuple = tuples'
-                alignTime = fetchn(stimulation.StimTrialEvents(tuple),'event_time');
-                endTime = fetch1(stimulation.StimTrialEvents(setfield(tuple,'event_type','endStimulus')),'event_time');
+                alignTime = double(fetchn(stimulation.StimTrialEvents(tuple),'event_time'));
+                endTime = double(fetch1(stimulation.StimTrialEvents(setfield(tuple,'event_type','endStimulus')),'event_time'));
                 
                 tuple.spikes_aligned = spikes(spikes > (alignTime - tuple.pre_stim_time) & ...
                     (spikes < (endTime + tuple.post_stim_time))) - alignTime;
